Clean up ・TileEntity系のクラスの名称変更 ・GUIのIDを定数化 ・不要なコメントアウト部分の削除
