+++ src/main/java/violetcraft/gui/ContainerViolet.java
@@ -4,11 +4,10 @@ import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.inventory.Container;
 import net.minecraft.inventory.Slot;
 import net.minecraft.item.ItemStack;
-import violetcraft.tiles.TileEntityGuiBlock;
+import violetcraft.tiles.TileGuiBlock;
 
 public class ContainerViolet extends Container
 {
-	private TileEntityGuiBlock tileEntity;
 	/** ブロック側のインベントリの第一スロットの番号 */
 	private static final int index0 = 0;
 	/** プレイヤーのインベントリの第一スロットの番号 */
@@ -17,8 +16,9 @@ public class ContainerViolet extends Container
 	private static final int index2 = 81;
 	/** このコンテナの全体のスロット数 */
 	private static final int index3 = 90;
+	private TileGuiBlock tileEntity;
 
-	public ContainerViolet(EntityPlayer player, TileEntityGuiBlock tileEntity) {
+	public ContainerViolet(EntityPlayer player, TileGuiBlock tileEntity) {
 		// スロットを設定する。
 		this.tileEntity = tileEntity;
 		for (int iy = 0; iy < 6; iy++) {
@@ -63,7 +63,7 @@ public class ContainerViolet extends Container
 			}
 
 			if (itemStack1.stackSize == 0) {
-				slot.putStack((ItemStack) null);
+				slot.putStack(null);
 			} else {
 				slot.onSlotChanged();
 			}
@@ -74,19 +74,4 @@ public class ContainerViolet extends Container
 		}
 		return itemStack;
 	}
-
-    //座標でGUIを開くか判定するためのもの。
-//    int xCoord, yCoord, zCoord;
-//    public ContainerViolet(int x, int y, int z) {
-//        this.xCoord = x;
-//        this.yCoord = y;
-//        this.zCoord = z;
-//    }
-
-//    @Override
-//    public boolean canInteractWith(EntityPlayer player) {
-//        もし、ブロックとの位置関係でGUI制御したいなら、こちらを使う
-//        return player.getDistanceSq(this.xCoord + 0.5D, this.yCoord + 0.5D, this.zCoord + 0.5D) <= 64D;
-//        return true;
-//    }
 }
