+++ src/main/java/violetcraft/gui/GuiHandler.java
@@ -4,8 +4,9 @@ import cpw.mods.fml.common.network.IGuiHandler;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.world.World;
-import violetcraft.tiles.TileEntityGenerator;
-import violetcraft.tiles.TileEntityGuiBlock;
+import violetcraft.VioletCraftMod;
+import violetcraft.tiles.TileGenerator;
+import violetcraft.tiles.TileGuiBlock;
 
 public class GuiHandler implements IGuiHandler
 {
@@ -17,16 +18,12 @@ public class GuiHandler implements IGuiHandler
 			return null;
 
     	TileEntity tileentity = world.getTileEntity(x, y, z);
-		if (tileentity instanceof TileEntityGuiBlock) {
-			return new ContainerViolet(player, (TileEntityGuiBlock) tileentity);
-		}
-        if (ID == 1) {
-        	return new ContainerGenerator(player, (TileEntityGenerator) tileentity);
+        if (tileentity instanceof TileGuiBlock) {
+            return new ContainerViolet(player, (TileGuiBlock) tileentity);
         }
-//		if (tileentity instanceof TileEntityGenerator) {
-//			return new ContainerGenerator(player, (TileEntityGenerator) tileentity);
-//		}
-
+        if (ID == VioletCraftMod.GUI_GENERATOR) {
+            return new ContainerGenerator(player, (TileGenerator) tileentity);
+        }
         return null;
     }
 
@@ -39,17 +36,13 @@ public class GuiHandler implements IGuiHandler
 
     	TileEntity tileentity = world.getTileEntity(x, y, z);
 
-    	if (tileentity instanceof TileEntityGuiBlock) {
-			return new GuiViolet(player, (TileEntityGuiBlock) tileentity);
-		}
-
-        if (ID == 1) {
-        	return new GuiGenerator(player, (TileEntityGenerator) tileentity);
+        if (tileentity instanceof TileGuiBlock) {
+            return new GuiViolet(player, (TileGuiBlock) tileentity);
         }
 
-//    	if (tileentity instanceof TileEntityGenerator) {
-//    		return new GuiGenerator(player, (TileEntityGenerator) tileentity);
-//    	}
+        if (ID == 1) {
+            return new GuiGenerator(player, (TileGenerator) tileentity);
+        }
     	return null;
